[Rename] modules/lang-expression
Refactor lang-expression module as part of the Elasticsearch to OpenSearch renaming.
