--- modules/lang-expression/src/main/java/org/elasticsearch/script/expression/ExpressionScoreScript.java
@@ -1,90 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.script.expression;
-
-import org.apache.lucene.expressions.Bindings;
-import org.apache.lucene.expressions.Expression;
-import org.apache.lucene.expressions.SimpleBindings;
-import org.apache.lucene.index.LeafReaderContext;
-import org.apache.lucene.search.DoubleValues;
-import org.apache.lucene.search.DoubleValuesSource;
-import org.elasticsearch.script.GeneralScriptException;
-import org.elasticsearch.script.ScoreScript;
-
-import java.io.IOException;
-
-/**
- * A bridge to evaluate an {@link Expression} against {@link Bindings} in the context
- * of a {@link ScoreScript}.
- */
-class ExpressionScoreScript implements ScoreScript.LeafFactory {
-
-    private final Expression exprScript;
-    private final DoubleValuesSource source;
-    private final boolean needsScores;
-
-    ExpressionScoreScript(Expression e, SimpleBindings b, boolean needsScores) {
-        this.exprScript = e;
-        this.source = exprScript.getDoubleValuesSource(b);
-        this.needsScores = needsScores;
-    }
-
-    @Override
-    public boolean needs_score() {
-        return needsScores;
-    }
-
-    @Override
-    public ScoreScript newInstance(final LeafReaderContext leaf) throws IOException {
-        return new ScoreScript(null, null, null) {
-            // Fake the scorer until setScorer is called.
-            DoubleValues values = source.getValues(leaf, new DoubleValues() {
-                @Override
-                public double doubleValue() throws IOException {
-                    return get_score();
-                }
-
-                @Override
-                public boolean advanceExact(int doc) throws IOException {
-                    return true;
-                }
-            });
-
-            @Override
-            public double execute(ExplanationHolder explanation) {
-                try {
-                    return values.doubleValue();
-                } catch (Exception exception) {
-                    throw new GeneralScriptException("Error evaluating " + exprScript, exception);
-                }
-            }
-
-            @Override
-            public void setDocument(int d) {
-                try {
-                    values.advanceExact(d);
-                } catch (IOException e) {
-                    throw new IllegalStateException("Can't advance to doc using " + exprScript, e);
-                }
-            }
-        };
-    }
-
-}
